pyproject.toml
[project]
name = "gamevault"
version = "0.1.0"
description = "Self-hosted game screenshot manager"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "Pillow>=11.0.0",
    "piexif>=1.1.3",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "markdown-it-py>=3.0.0",
    "Jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

# Note: Docker uses requirements.txt directly for faster builds.
# Keep requirements.txt in sync when updating dependencies above.

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
